Update README.md
### Additional Info about TBBT
The first episode of the show aired on September 24, 2007. The second pilot for the show was created. A different pilot was created for the 2006-07 television season, but it was never broadcast. The structure of the original unaired pilot differed from that of the current series. Leonard (Johnny Galecki) and Sheldon (Jim Parsons), who are named after Sheldon Leonard, a longtime figure in episodic television as a producer, director, and actor, were the only main characters retained in both pilots. Althea (Vernee Watson), a minor character, appeared in the first scene of both pilots and was kept largely intact.